+++ src/main/java/it/hurts/octostudios/rarcompat/items/feet/SnowshoesItem.java
@@ -36,7 +36,7 @@ public class SnowshoesItem extends WearableRelicItem {
                                 .stat(StatData.builder("amount")
                                         .initialValue(0.2D, 0.4D)
                                         .upgradeModifier(UpgradeOperation.MULTIPLY_BASE, 0.1D)
-                                        .formatValue(value -> MathUtils.round(value * 100, 1))
+                                        .formatValue(value -> (int) MathUtils.round(value * 100, 1))
                                         .build())
                                 .research(ResearchData.builder()
                                         .star(0, 6, 7).star(1, 11, 3).star(2, 16, 8).star(3, 15, 16)
