+++ src/main/java/it/hurts/octostudios/rarcompat/items/hat/DrinkingHatItem.java
@@ -34,7 +34,7 @@ public class DrinkingHatItem extends WearableRelicItem {
                                 .stat(StatData.builder("speed")
                                         .initialValue(0.3D, 0.4D)
                                         .upgradeModifier(UpgradeOperation.MULTIPLY_BASE, 0.15D)
-                                        .formatValue(value -> MathUtils.round(value * 100, 1))
+                                        .formatValue(value -> (int) MathUtils.round(value * 100, 1))
                                         .build())
                                 .research(ResearchData.builder()
                                         .star(0, 14, 4).star(1, 19, 9).star(2, 5, 11)
